foodgram_backend/requirements.txt
python-dotenv==0.20.0
Django==3.2
djangorestframework==3.12.4
django-colorfield==0.6.3
django-cors-headers==3.13.0
PyJWT==2.1.0
django-filter==23.3
psycopg2==2.9.3
flake8==5.0.4
gunicorn==20.1.0
webcolors==1.11.1
Pillow==9.0.0
djoser==2.1.0
drf-extra-fields==3.4.0
django-extensions

